pyproject.toml
[project]
name = "alertalot"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.37.0",
    "pyyaml>=6.0.2",
    "pytimeparse>=1.1.8",
    "jsonschema>=4.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pylint>=3.3.4",
]
